[#55] Provide a better error message in case of failure while cloning the repository (#56)
Fixes #55

Before this commit, the function that handled invoking the cloning command did not perform any error handling.
This PR attempts to fix that ; this means that `Shell.proc` now returns a `result` instead of `unit`, which is a breaking change. As such, the code that depends on this function - such as `clone_repo` in `Tui.Init` was updated to handle the result and exit with a better error message (well, at least one that is more useful).

Things that I am still not convinced about in this fix:
- The return type of `Shell.proc` is `(unit, string) result`. I am not really sure about crafting the error message there.
- `clone_repo` currently exits, but since it is an "internal" function (it is not exposed in the interface of the module), it is probably safer to make it return a `result` as well (as it is fallible) and let a public function exit on failure.